+++ server/src/main/java/org/opensearch/search/SearchBootstrapSettings.java
@@ -13,6 +13,8 @@ import org.opensearch.common.settings.Settings;
 
 /**
  * Keeps track of all the search related node level settings which can be accessed via static methods
+ *
+ * @opensearch.internal
  */
 public class SearchBootstrapSettings {
     // settings to configure maximum slice created per search request using OS custom slice computation mechanism. Default lucene
@@ -21,6 +23,7 @@ public class SearchBootstrapSettings {
     public static final int CONCURRENT_SEGMENT_SEARCH_TARGET_MAX_SLICE_COUNT_DEFAULT_VALUE = -1;
 
     // value <= 0 means lucene slice computation will be used
+    // this setting will be updated to dynamic setting as part of https://github.com/opensearch-project/OpenSearch/issues/8870
     public static final Setting<Integer> CONCURRENT_SEGMENT_SEARCH_TARGET_MAX_SLICE_COUNT_SETTING = Setting.intSetting(
         CONCURRENT_SEGMENT_SEARCH_TARGET_MAX_SLICE_COUNT_KEY,
         CONCURRENT_SEGMENT_SEARCH_TARGET_MAX_SLICE_COUNT_DEFAULT_VALUE,
@@ -32,7 +35,12 @@ public class SearchBootstrapSettings {
         settings = openSearchSettings;
     }
 
-    public static int getValueAsInt(String settingName, int defaultValue) {
-        return (settings != null) ? settings.getAsInt(settingName, defaultValue) : defaultValue;
+    public static int getTargetMaxSlice() {
+        return (settings != null)
+            ? settings.getAsInt(
+                CONCURRENT_SEGMENT_SEARCH_TARGET_MAX_SLICE_COUNT_KEY,
+                CONCURRENT_SEGMENT_SEARCH_TARGET_MAX_SLICE_COUNT_DEFAULT_VALUE
+            )
+            : CONCURRENT_SEGMENT_SEARCH_TARGET_MAX_SLICE_COUNT_DEFAULT_VALUE;
     }
 }